separate services to different files, remove redundant HPP code
