+++ .github/CODEOWNERS
@@ -1,1 +1,1 @@
-* @AlexandrosMor @rikterbeek @acampos1916 @candemiralp @michaelpaul @jillingk
+* @AlexandrosMor @rikterbeek @acampos1916 @candemiralp @michaelpaul @jillingk @antolo-arch
